requirements.txt
dnspython==1.16.0
ipykernel==6.29.5
mysql-connector-python==9.1.0
pip-chill==1.0.3
pymongo==3.11.0
python-dotenv==1.0.1
seaborn==0.13.2
sqlalchemy==2.0.36
